feat(camera): add touch input support for UI camera
Implement touch input handling for the UI camera in the `camera.rs` plugin. This
enables users to control the camera through touch gestures, such as panning and
pinching to zoom. The `InputHandlePlugin` is updated to include the new`camera_movement` system on `OnEnter(SceneState::GameScene)` event.

BREAKING CHANGE: The `UiCamera` component is now required to be added alongside
the `Camera2dBundle` in the `spawn_user_interface_camera` function.