feat(camera): make TouchTracker struct and fields public
The TouchTracker struct and its fields within the camera module are now
accessible from other modules due to the addition of the `pub` keyword. This
change is reflected in the `camera.rs` file.

Additionally, the `InputHandlePlugin` now inserts a default `TouchTracker`
resource into the application context, as seen in the `mod.rs` file.